+++ src/main/java/io/github/mmalykhin/hmsproxy/frontend/HortonworksFrontendExtension.java
@@ -4,4 +4,5 @@ public interface HortonworksFrontendExtension {
   Object addWriteNotificationLog(Object request) throws Throwable;
   Object getTablesExt(Object request) throws Throwable;
   Object getAllMaterializedViewObjectsForRewriting() throws Throwable;
+  Object set_aggr_stats_for(Object request) throws Throwable;
 }