+++ src/main/java/io/github/mmalykhin/hmsproxy/routing/RoutingHandler.java
@@ -156,6 +156,7 @@ final class RoutingHandler implements InvocationHandler, NamespaceFallback {
     SpecialCaseHandler getAllMvForRewriting = new GetAllMaterializedViewObjectsForRewritingHandler(support);
     SpecialCaseHandler lock = new LockRoutingHandler(support, this);
     SpecialCaseHandler refreshPrivileges = new RefreshPrivilegesHandler(support);
+    SpecialCaseHandler setAggrStatsFor = new SetAggrStatsForHandler(support);
     return Map.ofEntries(
         Map.entry("lock", lock),
         Map.entry("set_ugi", setUgi),
@@ -167,6 +168,7 @@ final class RoutingHandler implements InvocationHandler, NamespaceFallback {
         Map.entry("addWriteNotificationLog", addWriteNotificationLog),
         Map.entry("getTablesExt", getTablesExt),
         Map.entry("getAllMaterializedViewObjectsForRewriting", getAllMvForRewriting),
+        Map.entry("set_aggr_stats_for", setAggrStatsFor),
         Map.entry("drop_table", dropTable),
         Map.entry("drop_table_with_environment_context", dropTable),
         Map.entry("refresh_privileges", refreshPrivileges)